classes/sdcard_image-rpi: change to be usable with any image
Instead of being inherited by a specific image recipe, this class can
now be inherited globally in the machine configuration, allowing it to
be used with any image. This means we can remove "inherit image", and
we can also remove the IMAGE_FSTYPES modification since the machine
config does this already.

Signed-off-by: Paul Eggleton <paul.eggleton@linux.intel.com>