Use cmdline.txt to supply kernel command line
The build-time kernel command line (as defined by the CONFIG_CMDLINE config
parameter) is not respected by the RPI bootloader, and instead any custom
command line must be suplied via the cmdline.txt file in the boot partition.

This patch modifies the linux-raspberrypi recipe to generate a cmdline.txt
file that corresponds to the build-time configured command line, and it
also modifies the sdcard_image-rpi.bbclass to include this file in the
generated SD card images.

[GITHUB #62]