Use the softfp graphics libraries by default
By default the recipe will build everything with softfp optimization,
but the binary-only graphics libraries will be included as hardfp
by default. This results in non-working software (e.g. vchiq cannot
be called since it has the wrong ABI) and the OpenGL libraries don't
work. This patch switchs the default to the appropriate libraries.

Signed-off-by: Philipp Wagner <mail@philipp-wagner.com>