Add RPROVIDES for OpenGL libraries
Often the OpenGL libraries are dlopen'ed rather than linked against,
which makes it necessary to add OpenGL libraries to RDEPENDS of the
package. This of course requires that they are RPROVIDE'd somewhere.

I've seen a bug report about this at
https://github.com/djwillis/meta-raspberrypi/issues/55
but I don't know enough to judge how/why virtual/libgl is required,
so I'm leaving it in for now and only adding the runtime library
provider.

Signed-off-by: Philipp Wagner <mail@philipp-wagner.com>